Fixed bug with signup.

+++ db/schema.rb
@@ -22,7 +22,7 @@ ActiveRecord::Schema.define(version: 2022_02_18_195058) do
     t.boolean "complete"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
-    t.integer "user_id", null: false
+    t.bigint "user_id", null: false
     t.index ["user_id"], name: "index_tasks_on_user_id"
   end
